Fix YAML syntax (#20)
The example configuration will not propagate environment variables with
the current version of Serverless.

    plugins:
      serverless-sentry

parses as `{"plugins": "serverless-sentry"}`, instead of
`{"plugins": ["serverless-sentry"]}` - we mean the latter, so we need
a hyphen here. Perhaps Serverless used to allow a single string to
work as the value for the `plugins` key, but it does not in the current
version (1.26.1).