replace String with JSON.stringify
This is clearly a bug, because when I have value of sourceMaps in my config like
```
sourceMaps:
  urlPrefix: /var/task
```
It's serialized like `[object Object]` and that's what I see on environment variable in lambda configuration. I am not sure what this variables is needed for (is it needed at all?), but definitely it shouldn't look like that.

Could it be that for other cases here we should replace `String(...)` with `JSON.stringify(...)`?